+++ Soomla/Assets/Plugins/Soomla/Store/Config/StoreSettings.cs
@@ -73,7 +73,7 @@ namespace Soomla.Store
 		}
 
 		public void OnInfoGUI() {
-			SoomlaEditorScript.SelectableLabelField(frameworkVersion, "1.7.15");
+			SoomlaEditorScript.SelectableLabelField(frameworkVersion, "1.7.16");
 			SoomlaEditorScript.SelectableLabelField(buildVersion, "1");
 			EditorGUILayout.Space();
 		}