+++ Soomla/Assets/Plugins/Soomla/Store/SoomlaStore.cs
@@ -76,9 +76,9 @@ namespace Soomla.Store
 
 			StoreInfo.SetStoreAssets(storeAssets);
 
-            instance._loadBillingService();
-
-#if UNITY_IOS
+			instance._loadBillingService();
+			
+			#if UNITY_IOS
 			// On iOS we only refresh market items
 			instance._refreshMarketItemsDetails();
 #elif UNITY_ANDROID
@@ -168,10 +168,11 @@ namespace Soomla.Store
 			eventJSON.AddField("payload", payload);
 			StoreEvents.Instance.onMarketPurchaseStarted(eventJSON.print());
             
-			// in the editor we just give the item... no real market.
-			item.Give(1);
-            
-			// simulate onMarketPurchase event
+			// simulate events as they happen on the device
+			// the order is : 
+			//    onMarketPurchase
+			//    give item
+			//    onItemPurchase
 			StoreEvents.Instance.RunLater(() => {
 				eventJSON = new JSONObject();
 				eventJSON.AddField("itemId", item.ItemId);
@@ -186,6 +187,17 @@ namespace Soomla.Store
 			#endif
 				eventJSON.AddField("extra", extraJSON);
 				StoreEvents.Instance.onMarketPurchase(eventJSON.print());
+
+				// in the editor we just give the item... no real market.
+				item.Give(1);
+	
+				// We have to make sure the ItemPurchased event will be fired AFTER the balance/currency-changed events.
+				StoreEvents.Instance.RunLater(() => {
+					eventJSON = new JSONObject();
+					eventJSON.AddField("itemId", item.ItemId);
+					eventJSON.AddField("payload", payload);
+	            	StoreEvents.Instance.onItemPurchased(eventJSON.print());
+				});
 			});
 #endif
 		}