+++ soomla-native/projects/unity-wp8-fake-assembly/soomla-wp-core/soomla-wp-core/data/KeyValDatabase.cs
@@ -14,6 +14,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Threading;
 
 namespace SoomlaWpCore.data
 {