+++ soomla-native/projects/unity-wp8-fake-assembly/wp-store/wp-store/data/VirtualGoodsStorage.cs
@@ -18,7 +18,9 @@ using SoomlaWpCore.data;
 using SoomlaWpStore.domain;
 using SoomlaWpStore.domain.virtualGoods;
 using SoomlaWpStore;
+using SoomlaWpStore.events;
 using SoomlaWpStore.exceptions;
+using SoomlaWpCore.util;
 
 namespace SoomlaWpStore.data 
 {
@@ -60,7 +62,7 @@ public class VirtualGoodsStorage : VirtualItemStorage{
         KeyValueStorage.DeleteKeyValue(key);
 
         if (notify) {
-			StoreEvents.GetInstance().PostGoodUpgradeEvent(good,null);
+            BusProvider.Instance.Post(new GoodUpgradeEvent(good, null));
         }
     }
 
@@ -96,7 +98,7 @@ public class VirtualGoodsStorage : VirtualItemStorage{
         KeyValueStorage.SetValue(key, upItemId);
 
         if (notify) {
-			StoreEvents.GetInstance().PostGoodUpgradeEvent(good,upgradeVG);
+			BusProvider.Instance.Post(new GoodUpgradeEvent(good,upgradeVG));
         }
     }
 
@@ -205,7 +207,7 @@ public class VirtualGoodsStorage : VirtualItemStorage{
      * @{inheritDoc}
      */
     protected override void postBalanceChangeEvent(VirtualItem item, int balance, int amountAdded) {
-        StoreEvents.GetInstance().PostGoodBalanceChangedEvent((VirtualGood)item,balance, amountAdded);
+        BusProvider.Instance.Post(new GoodBalanceChangedEvent((VirtualGood)item,balance, amountAdded));
     }
 
     /**
@@ -220,12 +222,12 @@ public class VirtualGoodsStorage : VirtualItemStorage{
         if (equip) {
             KeyValueStorage.SetValue(key, "");
             if (notify) {
-                StoreEvents.GetInstance().PostGoodEquippedEvent(good);
+                BusProvider.Instance.Post(new GoodEquippedEvent(good));
             }
         } else {
             KeyValueStorage.DeleteKeyValue(key);
             if (notify) {
-                StoreEvents.GetInstance().PostGoodUnEquippedEvent(good);
+                BusProvider.Instance.Post(new GoodUnEquippedEvent(good));
             }
         }
     }