+++ soomla-native/projects/unity-wp8-fake-assembly/wp-store/wp-store/events/GoodEquippedEvent.cs
@@ -1,0 +1,26 @@
+﻿using SoomlaWpStore.domain.virtualGoods;
+using SoomlaWpCore.events;
+
+namespace SoomlaWpStore.events
+{
+    public class GoodEquippedEvent : SoomlaEvent
+    {
+        private EquippableVG mGood;
+
+        public EquippableVG GetEquippableVG()
+        {
+            return mGood;
+        }
+        
+        public GoodEquippedEvent(EquippableVG good)
+            : this(good, null)
+        {
+        }
+
+        public GoodEquippedEvent(EquippableVG good, object sender)
+            : base(sender)
+        {
+            mGood = good;
+        }
+    }
+}