+++ soomla-native/projects/unity-wp8-fake-assembly/wp-store/wp-store/events/UnexpectedStoreErrorEvent.cs
@@ -1,0 +1,25 @@
+﻿using SoomlaWpStore.domain;
+using SoomlaWpCore.events;
+
+namespace SoomlaWpStore.events
+{
+    public class UnexpectedStoreErrorEvent : SoomlaEvent
+    {
+        private string mMessage;
+        public string GetMessage()
+        {
+            return mMessage;
+        }
+
+        public UnexpectedStoreErrorEvent(string message)
+            : this(message,null)
+        {
+        }
+
+        public UnexpectedStoreErrorEvent(string message, object sender)
+            : base(sender)
+        {
+            mMessage = message;
+        }
+    }
+}