+++ soomla-native/projects/unity-wp8-fake-assembly/wp-store/wp-store/purchasesTypes/PurchaseWithVirtualItem.cs
@@ -15,7 +15,9 @@
 using System;
 using SoomlaWpStore;
 using SoomlaWpCore;
+using SoomlaWpCore.util;
 using SoomlaWpStore.data;
+using SoomlaWpStore.events;
 using SoomlaWpStore.domain;
 using SoomlaWpStore.exceptions;
 using System.Diagnostics;
@@ -61,7 +63,7 @@ public class PurchaseWithVirtualItem : PurchaseType {
             return;
         }
 
-		StoreEvents.GetInstance().PostItemPurchaseStartedEvent(getAssociatedItem());
+		BusProvider.Instance.Post(new ItemPurchaseStartedEvent(getAssociatedItem()));
 
         VirtualItemStorage storage = StorageManager.getVirtualItemStorage(item);
 
@@ -75,7 +77,7 @@ public class PurchaseWithVirtualItem : PurchaseType {
 
         getAssociatedItem().give(1);
         //BusProvider.getInstance().post(new OnItemPurchasedEvent(getAssociatedItem(), payload));
-		StoreEvents.GetInstance().PostItemPurchasedEvent(getAssociatedItem(), payload);
+		BusProvider.Instance.Post(new ItemPurchasedEvent(getAssociatedItem(), payload));
     }